Добавил новую сущность Category

+++ src/post/repository.ts
@@ -1,5 +1,6 @@
+import { AsyncResponse } from '../../core/response';
 import dataSource from '../../data-source';
-import { Post } from '../entities/post';
+import { Post } from '../../database/entities/post';
 
 export class PostRepository {
   private readonly repository;
@@ -8,29 +9,34 @@ export class PostRepository {
     this.repository = dataSource.getRepository(Post);
   }
 
-  public async one(id: number): Promise<any> {
+  public async one(id: number): AsyncResponse<Post> {
     return await this.repository.findOne({
+      relations: ['categories'],
       where: {
         id,
       },
     });
   }
 
-  public async all(): Promise<Post[]> {
-    return await this.repository.find();
+  public async all(): AsyncResponse<Post[]> {
+    return await this.repository.find({ relations: ['categories'] });
   }
 
-  public async create(data: Post): Promise<Post> {
-    return await this.repository.save(data);
+  public async create(data: Post): AsyncResponse<Post> {
+    const post = this.repository.create(data);
+
+    await this.repository.save(post);
+
+    return this.one(post.id);
   }
 
-  public async update(id: number, data: Post): Promise<Post> {
+  public async update(id: number, data: Post): AsyncResponse<Post> {
     await this.repository.update(id, data);
     return await this.one(id);
   }
 
-  public async remove(id: number): Promise<unknown> {
+  public async remove(id: number): AsyncResponse<unknown> {
     const currentPost = await this.one(id);
-    return await this.repository.remove(currentPost);
+    if (currentPost) return await this.repository.remove(currentPost);
   }
 }
